+++ radssh/config.py
@@ -1,5 +1,5 @@
 #
-# Copyright (c) 2014, 2016 LexisNexis Risk Data Management Inc.
+# Copyright (c) 2014, 2016, 2018 LexisNexis Risk Data Management Inc.
 #
 # This file is part of the RadSSH software package.
 #
@@ -50,6 +50,8 @@ shell.console=color
 # set this parameter to the number of lines to retain. Can be very memory
 # intensive.
 stalled_job_buffer=0
+# Report when command entered is modified (by a plugin) prior to executing
+show_altered_commands=off
 
 max_threads=120
 # Automatically save log files into date/time-stamped local directory